requirements.txt
pytest==8.0.0
requests==2.32.3
jsonschema==4.21.1
selene==2.0.0rc9
allure-pytest==2.13.5
allure-python-commons~=2.13.5
python-dotenv==1.0.1
